Change to using 'logging' subroutine directly from FTN::Log module.
Had been using a 'logged' subroutine in each script to do logging,
which in turn called the 'logging' subroutine in the FTN::Log
module to actually write out to the log.  This was left over from
when that 'logging' subroutine was actually inline instead of
in a separate module.  Change to calling the 'logging' subroutine
directly.