Strip EOL from nodelist lines in bin/nlftndb.
If the flags field is populated on a nodelist line (there is something after
the baud rate), then strip the EOL from the $flags string which holds the last
part of the line. It removes \r\n (windows) or \n (unix/linux) but not for \r
(like for a Classic MacOS text file).